Add files via upload
Download the monopoly.jpg file and change the path in the code to link to the path you have the image saved at. If you have Thorpy (popular GUI library used with Pygame) installed, then it should open up a window displaying the monopoly board when you run the program.